Fix eslint warnings
