+++ src/components/Circle.jsx
@@ -1,6 +1,5 @@
-import React from "react";
-import styled from "styled-components";
-
+import React from 'react';
+import styled from 'styled-components';
 
 const CircleStyle = styled.svg`
   fill: black;
@@ -9,9 +8,10 @@ const CircleStyle = styled.svg`
   margin-top: 0.54rem;
 `;
 
-export const Circle = () => (
-  <CircleStyle /* viewBox="0 0 20
-   20" */>
+const Circle = () => (
+  <CircleStyle>
     <path d="M4.4,8.8A4.4,4.4,0,1,1,8.8,4.4,4.41,4.41,0,0,1,4.4,8.8Z" />
   </CircleStyle>
 );
+
+export default Circle;
