+++ src/components/RenditionList.jsx
@@ -1,14 +1,14 @@
-import React, { useState, useRef, useEffect } from "react";
-import styled from "styled-components";
+import React, {useEffect, useRef, useState} from 'react';
+import styled from 'styled-components';
 
 const ListSection = styled.div`
-  padding: ${(props) => (props.mobile ? "0.9rem" : "2rem")};
+  padding: ${props => (props.mobile ? '0.9rem' : '2rem')};
   display: flex;
   flex-direction: column;
   background-color: var(--offwhite);
   margin-bottom: 1rem;
   &:last-child {
-    margin-bottom: ${(props) => props.mobile && "82px"};
+    margin-bottom: ${props => props.mobile && '82px'};
   }
   @media (max-width: 900px) {
     margin: 0 0.3rem 1rem 0.3rem;
@@ -33,42 +33,33 @@ const Title = styled.h3`
 const Details = styled.section`
   overflow: hidden;
   transition: max-height 0.3s ease;
-  max-height: ${(props) => props.height};
+  max-height: ${props => props.height};
 `;
 
 const PreviewDiv = styled.div`
-  height: ${(props) => props.height2};
+  height: ${props => props.height2};
   transition: 0.6s ease;
   overflow: hidden;
 `;
 
-const RenditionList = (props) => {
+const RenditionList = ({
+  description, descriptionPreview, expandValue, id, img, loaded, mobile, openAll, refClosedList, refOpenList, renditionsLength, title, year,
+}) => {
   const [active, setActive] = useState(false);
-  const [height, setHeight] = useState("0px");
-  const [height2, setHeight2] = useState("0px");
+  const [height, setHeight] = useState('0px');
+  const [height2, setHeight2] = useState('0px');
   /* const [propDependency, setPropDependency] = useState([props.renditionsLength, props.expandValue, props.loaded, props.id, active]) */
   const content = useRef(null);
   const content2 = useRef(null);
-
-  const {
-    renditionsLength,
-    expandValue,
-    loaded,
-    id,
-    refClosedList,
-    refOpenList,
-    openAll,
-    mobile,
-  } = props;
 
   useEffect(() => {
     refClosedList(content2.current.scrollHeight);
     refOpenList(content.current.scrollHeight);
     setActive(false);
     if (openAll || expandValue === id || mobile) setActive(true);
-    setHeight(active === false ? "0px" : `${content.current.scrollHeight}px`);
+    setHeight(active === false ? '0px' : `${content.current.scrollHeight}px`);
     setHeight2(
-      active === true && loaded ? "0px" : `${content2.current.scrollHeight}px`
+      active === true && loaded ? '0px' : `${content2.current.scrollHeight}px`,
     );
   }, [
     renditionsLength,
@@ -83,19 +74,19 @@ const RenditionList = (props) => {
   ]);
 
   return (
-    <ListSection mobile={props.mobile}>
+    <ListSection mobile={mobile}>
       <OpenImages>
         <TitleContainer>
-          <Title>{props.title}</Title>
-          <Title>{props.year}</Title>
+          <Title>{title}</Title>
+          <Title>{year}</Title>
         </TitleContainer>
-        <PreviewDiv aria-label="Open sections with buttons further down to read images" ref={content2} height2={height2}>
-          {props.descriptionPreview}
+        <PreviewDiv ref={content2} aria-label="Open sections with buttons further down to read images" height2={height2}>
+          {descriptionPreview}
         </PreviewDiv>
       </OpenImages>
       <Details ref={content} height={height}>
-        {props.img}
-        {props.description}
+        {img}
+        {description}
       </Details>
     </ListSection>
   );
