+++ src/pages/Home.jsx
@@ -1,0 +1,193 @@
+import React, {useState} from 'react';
+import {Helmet} from 'react-helmet';
+import Prismic from 'prismic-javascript';
+import styled from 'styled-components';
+import {useQuery} from 'react-query';
+
+import {apiEndpoint, client} from '../prismic-configuration';
+import GlobalStyle from '../styles/global';
+import Lines from '../components/Timeline';
+import Nav from '../components/Nav';
+import NotFound from './NotFound';
+
+const ua = navigator.userAgent;
+export const isMobile = /Android|webOS|iPhone|iPad|iPod/i.test(ua);
+
+const Main = styled.main`
+  margin: ${props => (props.mobile ? '0 5px 40px 5px' : '0 2rem 4rem 2rem')};
+  width: ${props => (props.mobile ? 'calc(100% - 10px)' : 'calc(100% - 4rem)')};
+  height: auto;
+  opacity: ${props => (props.loaded ? '1' : '0')};
+  transition: opacity 0.5s ease-in;
+`;
+const Loading = styled.p`
+  color: #fff;
+  margin: 32px 0 0 32px;
+  font-family: "PT-Regular", sans-serif;
+  font-size: 1.05rem;
+`;
+
+export const imgix = '&sat=-50&exp=0&invert=true&monochrome=c5c&con=-50&monochrome=%23862e9c';
+
+const Home = () => {
+  const [loaded, setLoaded] = useState(false);
+  const allLoaded = [];
+  const [email, setEmail] = useState(false);
+
+  const toggleTitle = () => {
+    setEmail(!email);
+  };
+
+  const getWorks = async () => {
+    try {
+      const works = await client.query(
+        Prismic.Predicates.at('document.type', 'work'),
+        {orderings: '[my.work.order, my.work.work_year_to desc]'},
+      );
+      return works;
+    }
+    catch {
+      throw new Error('No data found');
+    }
+  };
+
+  const createTimelineLinks = (data) => {
+    const works = data;
+    works.results.forEach((work, i) => {
+      const width = work.data.work_year_to - work.data.work_year_from + 1;
+      const newItem = {
+        ...work,
+        imageWidth: width,
+        link: {
+          id: work.id,
+          isBroken: false,
+          lang: work.lang,
+          link_type: 'Document',
+          tags: [],
+          type: work.type,
+          uid: work.uid,
+        },
+      };
+      works.results[i] = newItem;
+    });
+    return works;
+  };
+
+  const getMaxAndMinYears = (works) => {
+    let maxYear = works.results[0].data.work_year_to;
+    let minYear = works.results[0].data.work_year_from;
+    works.results.forEach((work) => {
+      minYear = work.data.work_year_from < minYear ? work.data.work_year_from : minYear;
+      maxYear = work.data.work_year_to > maxYear ? work.data.work_year_to : maxYear;
+    });
+    return {maxYear, minYear};
+  };
+
+  async function createTimelines() {
+    const works = await getWorks();
+    const worksWithLinkObject = {...createTimelineLinks(works)};
+    const {minYear, maxYear} = getMaxAndMinYears(worksWithLinkObject);
+    return {...worksWithLinkObject, maxYear, minYear};
+  }
+
+  const {
+    data: timelines, isError, isLoading, isSuccess,
+  } = useQuery('timelines', createTimelines);
+
+  const handleLoad = (i) => {
+    if (allLoaded.length === 0) {
+      timelines.results.forEach((work) => {
+        if (work.data.work_script.length) {
+          allLoaded.push(false);
+        }
+      });
+    }
+    allLoaded[i] = true;
+    if (allLoaded.every(value => value)) {
+      setLoaded(true);
+    }
+  };
+
+  if (isMobile && isSuccess) {
+    setTimeout(() => {
+      setLoaded(true);
+    }, 250);
+  }
+
+  if (isError) {
+    return <NotFound />;
+  }
+
+  if (isLoading) {
+    return (<Loading>Loading...</Loading>);
+  }
+
+  const repoNameArray = /([^/]+)\.cdn.prismic\.io\/api/.exec(apiEndpoint);
+  const repoName = repoNameArray[1];
+
+  return (
+    <>
+      <Helmet>
+        <title>Tris Vonna-Michell</title>
+        <meta content="Presentation of work by Tris Vonna-Michell." name="description" />
+        <meta
+          content="https://images.prismic.io/trisvonnamichell/6392235a-5597-4bb1-aa05-21a37c33b122_TVM-audio+poems-09+copy+2.jpg"
+          property="og:image"
+        />
+        <meta content="Presentation of work by Tris Vonna-Michell." property="og:description" />
+        <meta content="Tris Vonna-Michell" property="og:title" />
+        <meta content="summary_large_image" name="twitter:card" />
+        <meta content="Tris Vonna-Michell" name="twitter:title" />
+        <meta content="Presentation of work by Tris Vonna-Michell." name="twitter:description" />
+        <meta
+          content="https://images.prismic.io/trisvonnamichell/6392235a-5597-4bb1-aa05-21a37c33b122_TVM-audio+poems-09+copy+2.jpg"
+          name="twitter:image"
+        />
+        <link href="/favicon1.png" rel="shortcut icon" />
+        <script
+          async
+          defer
+          src={`//static.cdn.prismic.io/prismic.js?repo=${repoName}&new=true`}
+        />
+      </Helmet>
+      <Main loaded={loaded} mobile={isMobile}>
+        <GlobalStyle />
+        <Nav
+          mobile={isMobile}
+          title={!email ? 'Tris Vonna-Michell' : 'studiotvm@protonmail.com'}
+          toggleTitle={toggleTitle}
+          years={`Works ${timelines.min_year}–`}
+        />
+
+        {timelines.results.map((item) => {
+          const timelineWidth = timelines.max_year - timelines.min_year + 1;
+          return (
+            item.data.work_script.length > 0 && (
+            <Lines
+              key={item.link.id}
+              handleLoad={() => handleLoad(item.link.id)}
+              left={
+                ((item.data.work_year_from - timelines.min_year) / timelineWidth)
+                * 100
+              }
+              link={item.link}
+              loaded={loaded}
+              numberOfWorks={timelines.results.length}
+              renditions={false}
+              width={(item.image_width / timelineWidth) * 100}
+              workPreviewImage={item.data.work_preview_image.url}
+              workPreviewImageHeight={item.data.work_preview_image.dimensions.height}
+              workPreviewImageWidth={item.data.work_preview_image.dimensions.width}
+              workTitle={item.data.work_title[0].text}
+              workYearFrom={item.data.work_year_from}
+              workYearTo={item.data.work_year_to}
+            />
+            )
+          );
+        })}
+      </Main>
+    </>
+  );
+};
+
+export default Home;
