--- src/pages/index.js
@@ -1,3 +1,0 @@
-export { default as NotFound } from "./NotFound";
-export { default as Home } from "./Home";
-export { default as Renditions } from "./Renditions";
