+++ src/pages/index.jsx
@@ -1,0 +1,3 @@
+export {default as Home} from './Home';
+export {default as NotFound} from './NotFound';
+export {default as Renditions} from './Renditions';
