+++ src/prismic-configuration.js
@@ -1,30 +1,24 @@
-import Prismic from "prismic-javascript";
+import Prismic from 'prismic-javascript';
 
 // -- Prismic API endpoint
 // Determines which repository to query and fetch data from
-export const apiEndpoint = "https://trisvonnamichell.cdn.prismic.io/api/v2";
+export const apiEndpoint = 'https://trisvonnamichell.cdn.prismic.io/api/v2';
 
 // -- Access Token if the repository is not public
 // Generate a token in your dashboard and configure it here if your repository is private
-const accessToken = "";
+const accessToken = '';
 
 // -- Link resolution rules
 // Manages the url links to internal Prismic documents
 export const linkResolver = (doc) => {
-  if (doc.type === "work") return `/${doc.uid}`;
+  if (doc.type === 'work') return `/${doc.uid}`;
   // if (doc.type === "rendition") return `/${doc.uid}`;
 
-  return "/";
+  return '/';
 };
 
 // Client method to query documents from the Prismic repo
-export const client = Prismic.client(apiEndpoint, { accessToken });
-
-
-
-
-
-
+export const client = Prismic.client(apiEndpoint, {accessToken});
 
 /* // -- Link resolution rules
 // Manages the url links to internal Prismic documents
